rssreader/requirements.txt
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
attrs==23.1.0
autopep8==2.0.2
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.4.0
cryptography==40.0.2
Django==4.2.1
django-celery-beat==2.5.0
django-rest-knox==4.2.0
django-timezone-field==5.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
drf-spectacular-sidecar==2023.5.1
drf-standardized-errors==0.12.5
factory-boy==3.2.1
Faker==18.9.0
feedparser==6.0.10
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
kombu==5.2.4
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.26.1
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
uvloop==0.17.0
vine==5.0.0
watchdog==3.0.0
watchdog-gevent==0.1.1
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
zope.event==4.6
zope.interface==6.0
